Fix float parsing issue

+++ alt/fint.go
@@ -58,5 +58,5 @@ func ivalIntNotEmptyAsString(fi *finfo, rv reflect.Value, addr uintptr) (any, re
 	if v == 0 {
 		return nil, nilValue, true
 	}
-	return strconv.FormatInt(int64(v), 10), nilValue, false
+	return strconv.FormatInt(v, 10), nilValue, false
 }
